carbon-rag-service/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
langchain==1.2.10
langchain-community==0.4.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
faiss-cpu==1.11.0.post1
pypdf==5.9.0
docx2txt==0.9
